refactor a bunch of common code into one place
